requirements.txt
streamlit
pandas
plotly
openai
PyPDF2
python-docx
networkx